+++ R/visualization.R
@@ -209,9 +209,9 @@ objects are specified", call. = FALSE)
 #' @examples
 #' \dontshow{check_dependencies(python = TRUE) # make sure dependencies are present
 #' }
-#' # load an example model with an already simulated tree sequence and plot
-#' # it as an "admixture graph"
-#' model <- read_example("introgression")
+#' # load an example model with an already simulated tree sequence
+#' path <- system.file("extdata/models/introgression", package = "slendr")
+#' model <- read_model(path)
 #'
 #' plot_model(model, sizes = FALSE, log = TRUE)
 #' @export
@@ -440,11 +440,13 @@ plot_model <- function(model, sizes = TRUE, proportions = FALSE, log = FALSE) {
 #'
 #' @import ggplot2
 #' @export
-animate_model <- function(model, file = file.path(model$path, "output_ind_locations.tsv.gz"),
-                          steps, gif = NULL, width = 800, height = 560) {
+animate_model <- function(model, file, steps, gif = NULL, width = 800, height = 560) {
   if (!requireNamespace("magick", quietly = TRUE))
     message("For rendering animated GIFs, please install the R package ",
             "magick by calling `install.packages(\"magick\")")
+
+  if (!file.exists(file))
+    stop("Could not find file with locations at", file, call. = FALSE)
 
   if (!inherits(model$world, "slendr_map"))
     stop("Cannot animate non-spatial models", call. = FALSE)