+++ tests/testthat/helper-functions.R
@@ -9,11 +9,11 @@ run_sim <- function(pop, direction, simulation_length = NULL, method = "batch", 
     path = model_dir, direction = direction, simulation_length = simulation_length, overwrite = TRUE, force = TRUE
   )
 
-  slim(model, sequence_length = 1, recombination_rate = 0, save_locations = TRUE,
-       method = method, verbose = verbose)
+  locations_file <- tempfile(fileext = ".gz")
+  slim(model, sequence_length = 1, recombination_rate = 0,,
+       method = method, verbose = verbose, locations = locations_file)
 
-  df <- suppressMessages(readr::read_tsv(file.path(model$path, "output_ind_locations.tsv.gz"),
-                                         progress = FALSE)) %>%
+  df <- suppressMessages(readr::read_tsv(locations_file, progress = FALSE)) %>%
     dplyr::mutate(time = convert_slim_time(gen, model)) %>%
     dplyr::group_by(gen, time, pop) %>%
     dplyr::summarise(N = dplyr::n(), .groups = "keep")
@@ -31,28 +31,42 @@ run_sim <- function(pop, direction, simulation_length = NULL, method = "batch", 
 run_slim_msprime <- function(forward_model, backward_model,
                              forward_samples, backward_samples,
                              seq_len, rec_rate, seed, verbose) {
-  slim(forward_model, sequence_length = seq_len, recombination_rate = rec_rate,
-       sampling = forward_samples, random_seed = seed, verbose = verbose)
+  ts_slim_forward <- tempfile()
+  ts_msprime_forward <- tempfile()
+
+  slim(forward_model, output = ts_slim_forward, sequence_length = seq_len, recombination_rate = rec_rate,
+       samples = forward_samples, random_seed = seed, verbose = verbose)
   suppressWarnings({
-    msprime(forward_model, sequence_length = seq_len, recombination_rate = rec_rate,
-          sampling = forward_samples, random_seed = seed, verbose = verbose)
+    msprime(forward_model, output = ts_msprime_forward, sequence_length = seq_len, recombination_rate = rec_rate,
+          samples = forward_samples, random_seed = seed, verbose = verbose)
   })
 
-  slim(backward_model, sequence_length = seq_len, recombination_rate = rec_rate,
-       sampling = backward_samples, random_seed = seed, verbose = verbose)
+  ts_slim_backward <- tempfile()
+  ts_msprime_backward <- tempfile()
+
+  slim(backward_model, output = ts_slim_backward, sequence_length = seq_len, recombination_rate = rec_rate,
+       samples = backward_samples, random_seed = seed, verbose = verbose)
   suppressWarnings({
-  msprime(backward_model, sequence_length = seq_len, recombination_rate = rec_rate,
-          sampling = backward_samples, random_seed = seed, verbose = verbose)
+  msprime(backward_model, output = ts_msprime_backward, sequence_length = seq_len, recombination_rate = rec_rate,
+          samples = backward_samples, random_seed = seed, verbose = verbose)
   })
+
+  list(
+    "slim_forward"     = ts_slim_forward,
+    "msprime_forward"  = ts_msprime_forward,
+    "slim_backward"    = ts_slim_backward,
+    "msprime_backward" = ts_msprime_backward
+  )
 }
 
-load_tree_sequence <- function(backend, model, N, rec_rate, mut_rate, seed) {
-  if (backend == "SLiM")
+load_tree_sequence <- function(backend, direction, ts_list, model, N, rec_rate, mut_rate, seed) {
+  ts_file <- ts_list[[paste(tolower(backend), direction, sep = "_")]]
+  if (backend == tolower("SLiM"))
     ts_load(
-      model, file = file.path(model$path, "output_slim.trees"), recapitate = TRUE, simplify = TRUE, mutate = TRUE,
+      model = model, file = ts_file, recapitate = TRUE, simplify = TRUE, mutate = TRUE,
       Ne = N, recombination_rate = rec_rate, mutation_rate = mut_rate,
       random_seed = seed
     )
   else
-    ts_load(model, file = file.path(model$path, "output_msprime.trees"), mutate = TRUE, mutation_rate = mut_rate, random_seed = seed)
+    ts_load(model = model, file = ts_file, mutate = TRUE, mutation_rate = mut_rate, random_seed = seed)
 }